airflow_playground/requirements.txt
psycopg2-binary
apache-airflow-providers-clickhouse
